update: relocate testHelper
